requirements.txt
flask>=2.3.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
faiss-cpu>=1.7.0  # 或 faiss-gpu 如果需要GPU
scikit-learn>=1.3.0